+++ test/test_yang_deviation.sh
@@ -35,11 +35,15 @@ module example-base{
    prefix base;
    namespace "urn:example:base";
    container system {
-      must "daytime or time";
-      leaf daytime{ 	    /* not supported removes this */
+      must "daytime or time"; /* deviate delete removes this */
+      leaf daytime{ 	      /* deviate not-supported removes this */
+         type string;
+      }
+      leaf time{ 	    
          type string;
       }
       list name-server {
+         max-elements 1;      /* deviate replace replaces to "max.elements 3" here */
          key name;
          leaf name {
             type string;
@@ -52,7 +56,7 @@ module example-base{
          } 
          leaf type {
             type string;
-	    /* add rule adds default here */
+	    /* deviate add adds "default admin" here */
          } 
       }
    }
@@ -60,13 +64,17 @@ module example-base{
 EOF
 
 # Args:
-# 0: daytime implemented: true/false
-# 1: admin type default: true/false
+# 1: daytime implemented: true/false
+# 2: admin type default: true/false
+# 3: mustdate default: true/false
+# 4: maxelement of name-server is 1: true/false (if false the # is 3)
 function testrun()
 {
     daytime=$1
     admindefault=$2
-    
+    mustdate=$3
+    maxel1=$4
+
     new "test params: -f $cfg"
 
     if [ "$BE" -ne 0 ]; then
@@ -82,29 +90,55 @@ function testrun()
     new "wait backend"
     wait_backend
 
-    if ! $daytime; then # Not supported - dont continue
-	new "Add example-base daytime - should not be supported"
-	expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><edit-config><target><candidate/></target><config><system xmlns=\"urn:example:base\"><daytime>Sept17</daytime></system></config></edit-config></rpc>]]>]]>" "<rpc-reply $DEFAULTNS><rpc-error><error-type>application</error-type><error-tag>unknown-element</error-tag><error-info><bad-element>daytime</bad-element></error-info><error-severity>error</error-severity><error-message>Failed to find YANG spec of XML node: daytime with parent: system in namespace: urn:example:base</error-message></rpc-error></rpc-reply>]]>]]"
+    new "Add user bob"
+    expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><edit-config><target><candidate/></target><config><system xmlns=\"urn:example:base\"><user><name>bob</name></user></system></config></edit-config></rpc>]]>]]>" "<rpc-reply $DEFAULTNS><ok/></rpc-reply>]]>]]"
+
+    if $mustdate; then # fail since there is neither date or daytime (delete rule)
+	new "netconf validate expect error"
+	expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><validate><source><candidate/></source></validate></rpc>]]>]]>" "^<rpc-reply $DEFAULTNS><rpc-error><error-type>application</error-type><error-tag>operation-failed</error-tag><error-severity>error</error-severity><error-message>must xpath validation failed</error-message></rpc-error></rpc-reply>]]>]]>$"
     else
+	new "netconf validate ok"
+	expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><validate><source><candidate/></source></validate></rpc>]]>]]>" "<rpc-reply $DEFAULTNS><ok/></rpc-reply>]]>]]"
+    fi
+
+    new "Add time"
+    expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><edit-config><target><candidate/></target><config><system xmlns=\"urn:example:base\"><time>yes</time></system></config></edit-config></rpc>]]>]]>" "<rpc-reply $DEFAULTNS><ok/></rpc-reply>]]>]]"
+
+    new "netconf validate ok"
+    expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><validate><source><candidate/></source></validate></rpc>]]>]]>" "<rpc-reply $DEFAULTNS><ok/></rpc-reply>]]>]]"
+
+    if $daytime; then # not-supported rule
 	new "Add example-base daytime - supported"
 	expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><edit-config><target><candidate/></target><config><system xmlns=\"urn:example:base\"><daytime>Sept17</daytime></system></config></edit-config></rpc>]]>]]>" "<rpc-reply $DEFAULTNS><ok/></rpc-reply>]]>]]"
+    else # Not supported
+	new "Add example-base daytime - expect error not supported"
+	expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><edit-config><target><candidate/></target><config><system xmlns=\"urn:example:base\"><daytime>Sept17</daytime></system></config></edit-config></rpc>]]>]]>" "<rpc-reply $DEFAULTNS><rpc-error><error-type>application</error-type><error-tag>unknown-element</error-tag><error-info><bad-element>daytime</bad-element></error-info><error-severity>error</error-severity><error-message>Failed to find YANG spec of XML node: daytime with parent: system in namespace: urn:example:base</error-message></rpc-error></rpc-reply>]]>]]"
+    fi # daytime supported
 
-	new "Add user bob"
-	expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><edit-config><target><candidate/></target><config><system xmlns=\"urn:example:base\"><user><name>bob</name></user></system></config></edit-config></rpc>]]>]]>" "<rpc-reply $DEFAULTNS><ok/></rpc-reply>]]>]]"
-
-	new "netconf commit"
-	expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><commit/></rpc>]]>]]>" "^<rpc-reply $DEFAULTNS><ok/></rpc-reply>]]>]]>$"
+    new "netconf commit"
+    expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><commit/></rpc>]]>]]>" "^<rpc-reply $DEFAULTNS><ok/></rpc-reply>]]>]]>$"
 	
-	if $admindefault; then 
-	    new "Get type admin expected"
-	    expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><get-config><source><running/></source><filter type=\"xpath\" select=\"/base:system/base:user[base:name='bob']\" xmlns:base=\"urn:example:base\"/></get-config></rpc>]]>]]>" "^<rpc-reply $DEFAULTNS><data><system xmlns=\"urn:example:base\"><user><name>bob</name><type>admin</type></user></system></data></rpc-reply>]]>]]>$"
+    if $admindefault; then # add rule
+	new "Get type admin expected"
+	expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><get-config><source><running/></source><filter type=\"xpath\" select=\"/base:system/base:user[base:name='bob']\" xmlns:base=\"urn:example:base\"/></get-config></rpc>]]>]]>" "^<rpc-reply $DEFAULTNS><data><system xmlns=\"urn:example:base\"><user><name>bob</name><type>admin</type></user></system></data></rpc-reply>]]>]]>$"
 # XXX Cannot select a default value??	    
-#	    expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><get-config><source><running/></source><filter type=\"xpath\" select=\"/base:system/base:user[base:name='bob']/base:type\" xmlns:base=\"urn:example:base\"/></get-config></rpc>]]>]]>" foo
-	else
-	    new "Get type none expected"
-	    expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><get-config><source><running/></source><filter type=\"xpath\" select=\"/base:system/base:user[base:name='bob']/base:type\" xmlns:base=\"urn:example:base\"/></get-config></rpc>]]>]]>" "^<rpc-reply $DEFAULTNS><data/></rpc-reply>]]>]]>$"
-	fi
+	#	    expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><get-config><source><running/></source><filter type=\"xpath\" select=\"/base:system/base:user[base:name='bob']/base:type\" xmlns:base=\"urn:example:base\"/></get-config></rpc>]]>]]>" foo
+    else
+	new "Get type none expected"
+	expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><get-config><source><running/></source><filter type=\"xpath\" select=\"/base:system/base:user[base:name='bob']/base:type\" xmlns:base=\"urn:example:base\"/></get-config></rpc>]]>]]>" "^<rpc-reply $DEFAULTNS><data/></rpc-reply>]]>]]>$"
     fi
+
+    # Add 2 name-servers
+    new "Add two name-servers"
+    expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><edit-config><target><candidate/></target><config><system xmlns=\"urn:example:base\"><name-server><name>aa</name></name-server><name-server><name>bb</name></name-server></system></config></edit-config></rpc>]]>]]>" "<rpc-reply $DEFAULTNS><ok/></rpc-reply>]]>]]"
+    if $maxel1; then # add two and check if it fails
+	new "netconf validate 2 element fail"
+	expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><validate><source><candidate/></source></validate></rpc>]]>]]>" "^<rpc-reply $DEFAULTNS><rpc-error><error-type>protocol</error-type><error-tag>operation-failed</error-tag><error-app-tag>too-many-elements</error-app-tag><error-severity>error</error-severity><error-path>/system/name-server</error-path></rpc-error></rpc-reply>]]>]]>$"
+    else
+	new "netconf validate 2 elements ok"
+	expecteof "$clixon_netconf -qf $cfg" 0 "$DEFAULTHELLO<rpc $DEFAULTNS><validate><source><candidate/></source></validate></rpc>]]>]]>" "<rpc-reply $DEFAULTNS><ok/></rpc-reply>]]>]]"
+    fi
+
     if [ "$BE" -ne 0 ]; then
 	new "Kill backend"
 	# Check if premature kill
@@ -128,8 +162,8 @@ module example-deviations{
    }
 }
 EOF
-new "daytime supported"
-testrun true false
+new "1. daytime supported"
+testrun true false true true
 
 # Example from RFC 7950 Sec 7.20.3.3
 cat <<EOF > $fyangdev
@@ -145,10 +179,10 @@ module example-deviations{
    }
 }
 EOF
-new "daytime not supported"
-testrun false false
+new "2. daytime not supported"
+testrun false false true true
 
-# Example from RFC 7950 Sec 7.20.3.3
+# Add example from RFC 7950 Sec 7.20.3.3
 cat <<EOF > $fyangdev
 module example-deviations{
    yang-version 1.1;
@@ -164,8 +198,47 @@ module example-deviations{
    }
 }
 EOF
-new "deviate add, check admin default"
-testrun true true
+new "3. deviate add, check admin default"
+testrun true true true true
+
+# Delete example from RFC 7950 Sec 7.20.3.3
+cat <<EOF > $fyangdev
+module example-deviations{
+   yang-version 1.1;
+   prefix md;
+   namespace "urn:example:deviations";
+   import example-base {
+         prefix base;
+   }
+   deviation /base:system/base:name-server {
+      deviate replace {
+         max-elements 3;
+      }
+   }
+}
+EOF
+new "4. deviate replace"
+testrun true false true false
+
+# Replace example from RFC 7950 Sec 7.20.3.3
+cat <<EOF > $fyangdev
+module example-deviations{
+   yang-version 1.1;
+   prefix md;
+   namespace "urn:example:deviations";
+   import example-base {
+         prefix base;
+   }
+   deviation /base:system {
+      deviate delete {
+         must "daytime or time";
+      }
+   }
+}
+EOF
+
+new "5. deviate delete"
+testrun true false false true
 
 rm -rf "$dir"